refactor: move tests in tests.rs
